Create Main Script
This script will ask for an input of the user's logo name, as well as their workstations hostname and store them in variables. It will then generate a random password and change the user's password to that. After, the user will be removed from all AD groups and have their account as well as their workstation disabled in AD. Finally, a log file will be generated on your desktop, providing you confirmation on what has occurred.